Extended interpreter to understand floating point vs interger numbers.
Grammar Changes:

term                 :: factor ((MUL | INTEGER_DIV | FLOAT_DIV) factor)*
factor               :: (PLUS | MINUS) factor | INTEGER_CONST | REAL_CONST | LPAREN expr RPAREN | variable

- Added INTEGER_DIV, FLOAT_DIV, INTEGER_CONST, and REAL_CONST tokens
- Removed DIVIDE and INTEGER tokens
- Updated lexer to handle new tokens
- Updated parser/ast to handle grammar changes
- Updated interpreter to handle floating point vs. integer division

+++ src/interpreter/node_visitor.rs
@@ -8,11 +8,11 @@ use parser::ast::Variable;
 
 pub trait NodeVisitor {
 
-    fn visit_variable(&mut self, node: Variable) -> Result<i32, String>;
-    fn visit_unaryop(&mut self, node: Expr) -> Result<i32, String>;
-    fn visit_binop(&mut self, node: Expr) -> Result<i32, String>;
-    fn visit_num(&mut self, node: Expr) -> Result<i32, String>;
-    fn visit_expr(&mut self, node: Expr) -> Result<i32, String> {
+    fn visit_variable(&mut self, node: Variable) -> Result<f32, String>;
+    fn visit_unaryop(&mut self, node: Expr) -> Result<f32, String>;
+    fn visit_binop(&mut self, node: Expr) -> Result<f32, String>;
+    fn visit_num(&mut self, node: Expr) -> Result<f32, String>;
+    fn visit_expr(&mut self, node: Expr) -> Result<f32, String> {
         return match node {
             Expr::UnaryOp(_, _)      => self.visit_unaryop(node),
             Expr::BinOp(_, _, _)     => self.visit_binop(node),
